refactor: merge dockerexec.py into podman.py
Move ExecSession class from dockerexec.py into podman.py so all
podman CLI interactions live in one module. Update imports in
wshandler.py and rename test_dockerexec.py to test_podman_exec.py.

Closes #143
